requirements.txt
amqp==5.0.3
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.5
django-background-tasks==1.2.5
django-celery-results==2.0.1
django-compat==1.0.15
idna==2.10
kombu==5.0.2
prompt-toolkit==3.0.14
psycopg2-binary==2.8.6
pytz==2020.5
redis==3.5.3
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
